Move files into correct package
