+++ config/checkstyle/checkstyle.xml
@@ -135,7 +135,7 @@
         </module>
         <module name="PackageName">
             <property name="format"
-                      value="^[a-z]+(\.[a-z][a-z0-9_]*[a-z0-9]+)*$" />
+                      value="^[a-z]+(\.[a-z][a-z0-9_]*[a-z0-9O]+)*$" />
             <message key="name.invalidPattern"
                      value="Package name ''{0}'' must match pattern ''{1}''." />
         </module>
