+++ src/test/java/edu/wpi/cs3733/d19/teamO/UtilityClassTest.java
@@ -1,4 +1,4 @@
-package edu.wpi.cs3733d18.onyx_owlmen;
+package edu.wpi.cs3733.d19.teamO;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
